+++ app/src/main/java/woowacourse/shopping/view/productDetail/ProductDetailActivity.kt
@@ -1,0 +1,67 @@
+package woowacourse.shopping.view.productDetail
+
+import android.content.Context
+import android.content.Intent
+import android.os.Build
+import android.os.Bundle
+import android.widget.Toast
+import androidx.activity.enableEdgeToEdge
+import androidx.activity.viewModels
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import woowacourse.shopping.R
+import woowacourse.shopping.databinding.ActivityProductDetailBinding
+import woowacourse.shopping.domain.product.Product
+
+class ProductDetailActivity : AppCompatActivity() {
+    private val viewModel: ProductDetailViewModel by viewModels()
+    private val binding: ActivityProductDetailBinding by lazy {
+        ActivityProductDetailBinding.inflate(layoutInflater)
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContentView(binding.root)
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.productDetailRoot)) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            insets
+        }
+
+        initViewModel()
+        bindViewModel()
+    }
+
+    private fun initViewModel() {
+        val product = intent.getProductExtra() ?: run {
+            Toast.makeText(this, "상품 정보를 불러오지 못했습니다.", Toast.LENGTH_SHORT).show()
+            return finish()
+        }
+        viewModel.updateProduct(product)
+    }
+
+    private fun bindViewModel() {
+        binding.viewModel = viewModel
+        binding.lifecycleOwner = this
+    }
+
+    private fun Intent.getProductExtra(): Product? {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            getSerializableExtra(EXTRA_PRODUCT, Product::class.java)
+        } else {
+            getSerializableExtra(EXTRA_PRODUCT) as? Product
+        }
+    }
+
+    companion object {
+        private const val EXTRA_PRODUCT = "woowacourse.shopping.EXTRA_PRODUCT"
+
+        fun newIntent(context: Context, product: Product): Intent {
+            return Intent(context, ProductDetailActivity::class.java).putExtra(
+                EXTRA_PRODUCT, product
+            )
+        }
+    }
+}
